Creating an Atom syntax theme based on a Vim theme by github.com/rakr
I really like how rakr combined a duotone theme with the Panic Firewatch
colors to create a light and dark syntax theme for Vim. I wanted the
light theme in Atom, so I decided to learn how to create an Atom syntax
theme. I used a duotone theme by github.com/smlombardi as a model/
template for my theme and then swapped in the colors from rakr's Vim
theme.

I still need to add and/or improve some of the language support. In
particular, I want a better syntax coloring for Pweave Markdown files
and for Bash script files.

 On branch master

 Initial commit

	new file:   .gitignore
	new file:   CHANGELOG.md
	new file:   LICENSE.md
	new file:   README.md
	new file:   index.less
	new file:   package.json
	new file:   styles/colors.less
	new file:   styles/editor.less
	new file:   styles/languages/_base.less
	new file:   styles/languages/_index.less
	new file:   styles/languages/c.less
	new file:   styles/languages/clojure.less
	new file:   styles/languages/coffee.less
	new file:   styles/languages/cs.less
	new file:   styles/languages/css.less
	new file:   styles/languages/gfm.less
	new file:   styles/languages/haskell.less
	new file:   styles/languages/html.less
	new file:   styles/languages/jade.less
	new file:   styles/languages/java.less
	new file:   styles/languages/javascript.less
	new file:   styles/languages/json.less
	new file:   styles/languages/less.less
	new file:   styles/languages/mediawiki.less
	new file:   styles/languages/php.less
	new file:   styles/languages/pmd.less
	new file:   styles/languages/python.less
	new file:   styles/languages/sass.less
	new file:   styles/languages/scss.less
	new file:   styles/languages/slim.less
	new file:   styles/languages/stylus.less
	new file:   styles/languages/tex.less
	new file:   styles/languages/text.less
	new file:   styles/languages/yaml.less
	new file:   styles/syntax-variables.less